requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.0.1
torchaudio==2.0.2
numpy==1.24.3
tqdm==4.66.1